refactor(core): address System Record ownership review
