+++ packages/core/package.json
@@ -28,7 +28,7 @@
     "test": "pnpm run test:baseline && pnpm run test:system-record",
     "test:baseline": "vitest run --exclude 'test/system-record-*.test.ts' --maxWorkers=4",
     "test:system-record": "vitest run --config vitest.system-record.config.ts --maxWorkers=1 --no-file-parallelism",
-    "test:system-record-export": "node test/system-record-package-export-v1.mjs",
+    "test:system-record-export": "node test/system-record-package-export-v1.mjs && tsc --noEmit --strict --skipLibCheck --target ES2022 --module NodeNext --moduleResolution NodeNext test/system-record-package-export-v1.types.ts",
     "test:system-record-exhaustive": "DKG_SYSTEM_RECORD_EXHAUSTIVE=1 vitest run test/system-record-inventory-v1.test.ts test/system-record-objects-v1.test.ts -t \"complete height-three|exactly 128 MiB\" --maxWorkers=1 --no-file-parallelism",
     "test:coverage": "vitest run --coverage",
     "clean": "rm -rf dist tsconfig.tsbuildinfo"
