+++ packages/core/src/system-record-objects-v1.ts
@@ -6,73 +6,85 @@
  * factory-only capabilities and orchestration helpers from becoming public API.
  */
 export {
-  AGENT_PROFILE_LINK_PREDICATES_V1,
-  AGENT_PROFILE_SCHEMA_TERMS_V1,
   assertAgentProfileAuthorityTransitionV1,
-  assertAgentProfileConflictEvidenceV1,
   assertAgentProfileForkResolutionV1,
-  assertAgentProfileHeadObjectV1,
-  assertAgentRootV1,
-  assertCanonicalRfc3339SecondsV1,
-  assertCanonicalSystemRecordPeerIdV1,
-  assertDerivedAgentEncryptionSubjectV1,
-  assertOwnedSubjectTableObjectV1,
-  assertSystemRecordPeerBindingV1,
   canonicalizeAgentProfileAuthorityTransitionV1,
-  canonicalizeAgentProfileConflictEvidenceV1,
   canonicalizeAgentProfileForkResolutionV1,
-  canonicalizeAgentProfileHeadObjectV1,
-  canonicalizeOwnedSubjectTableObjectV1,
-  canonicalizeSystemRecordRootCollisionEvidenceV1,
-  classifyAgentProfileOwnedSubjectV1,
   computeAgentProfileAuthorityTransitionDigestV1,
-  computeAgentProfileConflictEvidenceDigestV1,
   computeAgentProfileForkResolutionDigestV1,
-  computeAgentProfileHeadObjectDigestV1,
-  computeOwnedSubjectTableDigestV1,
-  computeSystemRecordRootCollisionEvidenceDigestV1,
-  copyBoundedSystemRecordBytesV1,
-  decodeUnpaddedBase64UrlV1,
-  digestSystemRecordJsonV1,
-  digestSystemRecordBytesV1,
-  deriveAgentProfileOwnedSubjectV1,
-  EMPTY_OWNED_SUBJECT_TABLE_DIGEST_V1,
-  isAllowedAgentProfilePredicateV1,
   parseCanonicalAgentProfileAuthorityTransitionV1,
-  parseCanonicalAgentProfileConflictEvidenceV1,
   parseCanonicalAgentProfileForkResolutionV1,
-  parseCanonicalAgentProfileHeadObjectV1,
-  parseCanonicalOwnedSubjectTableObjectV1,
-  parseCanonicalRfc3339SecondsV1,
-  SystemRecordObjectErrorV1,
-  type AgentProfileActiveHeadObjectV1,
-  type AgentProfileExactLinkedSubjectKindV1,
   type AgentProfileAuthorityTransitionV1,
-  type AgentProfileConflictEvidenceV1,
-  type AgentProfileForkConflictEntryV1,
   type AgentProfileForkResolutionV1,
-  type AgentProfileHeadCommonV1,
-  type AgentProfileHeadObjectV1,
-  type AgentProfileIndexedSubjectKindV1,
-  type AgentProfileLinkedSubjectKindV1,
-  type AgentProfileOwnedSubjectKindV1,
-  type AgentProfileTombstoneHeadObjectV1,
-  type AgentProfileTransitionConflictEntryV1,
-  type CanonicalRfc3339SecondsV1,
-  type OwnedSubjectTableObjectV1,
   type SignedAgentProfileAuthorityTransitionEnvelopeV1,
   type SignedAgentProfileForkResolutionEnvelopeV1,
   type SignedAgentProfileHeadEnvelopeV1,
   type SignedSystemRecordEnvelopeV1,
   type SystemRecordEip1271EvidenceV1,
   type SystemRecordNoSignatureEvidenceV1,
-  type SystemRecordObjectErrorCodeV1,
-  type SystemRecordPeerPublicKeyV1,
-  type SystemRecordRootCollisionEvidenceV1,
   type SystemRecordSignatureEntryV1,
   type SystemRecordSignatureRoleV1,
   type SystemRecordSignatureSuiteV1,
-} from './system-record-agent-profile-codecs-v1-internal.js';
+} from './system-record-agent-profile-control-codecs-v1-internal.js';
+
+export {
+  assertAgentProfileHeadObjectV1,
+  canonicalizeAgentProfileHeadObjectV1,
+  computeAgentProfileHeadObjectDigestV1,
+  parseCanonicalAgentProfileHeadObjectV1,
+  type AgentProfileActiveHeadObjectV1,
+  type AgentProfileHeadCommonV1,
+  type AgentProfileHeadObjectV1,
+  type AgentProfileTombstoneHeadObjectV1,
+} from './system-record-agent-profile-head-codec-v1-internal.js';
+
+export {
+  assertAgentRootV1,
+  assertCanonicalRfc3339SecondsV1,
+  assertCanonicalSystemRecordPeerIdV1,
+  assertSystemRecordPeerBindingV1,
+  copyBoundedSystemRecordBytesV1,
+  decodeUnpaddedBase64UrlV1,
+  digestSystemRecordBytesV1,
+  digestSystemRecordJsonV1,
+  parseCanonicalRfc3339SecondsV1,
+  SystemRecordObjectErrorV1,
+  type CanonicalRfc3339SecondsV1,
+  type SystemRecordObjectErrorCodeV1,
+  type SystemRecordPeerPublicKeyV1,
+} from './system-record-agent-profile-primitives-v1-internal.js';
+
+export {
+  assertAgentProfileConflictEvidenceV1,
+  canonicalizeAgentProfileConflictEvidenceV1,
+  canonicalizeSystemRecordRootCollisionEvidenceV1,
+  computeAgentProfileConflictEvidenceDigestV1,
+  computeSystemRecordRootCollisionEvidenceDigestV1,
+  parseCanonicalAgentProfileConflictEvidenceV1,
+  type AgentProfileConflictEvidenceV1,
+  type AgentProfileForkConflictEntryV1,
+  type AgentProfileTransitionConflictEntryV1,
+  type SystemRecordRootCollisionEvidenceV1,
+} from './system-record-agent-profile-evidence-codecs-v1-internal.js';
+
+export {
+  AGENT_PROFILE_LINK_PREDICATES_V1,
+  AGENT_PROFILE_SCHEMA_TERMS_V1,
+  assertDerivedAgentEncryptionSubjectV1,
+  assertOwnedSubjectTableObjectV1,
+  canonicalizeOwnedSubjectTableObjectV1,
+  classifyAgentProfileOwnedSubjectV1,
+  computeOwnedSubjectTableDigestV1,
+  deriveAgentProfileOwnedSubjectV1,
+  EMPTY_OWNED_SUBJECT_TABLE_DIGEST_V1,
+  isAllowedAgentProfilePredicateV1,
+  parseCanonicalOwnedSubjectTableObjectV1,
+  type AgentProfileExactLinkedSubjectKindV1,
+  type AgentProfileIndexedSubjectKindV1,
+  type AgentProfileLinkedSubjectKindV1,
+  type AgentProfileOwnedSubjectKindV1,
+  type OwnedSubjectTableObjectV1,
+} from './system-record-owned-subject-codecs-v1-internal.js';
 
 export {
   assertCanonicalEip191SignatureV1,
