+++ packages/core/test/system-record-module-boundaries-v1.test.ts
@@ -6,19 +6,34 @@ const source = (name: string): string =>
   readFileSync(new URL(`../src/${name}`, import.meta.url), 'utf8');
 
 describe('System Record V1 module ownership', () => {
-  it('keeps the supported package, object, and inventory facades explicit', () => {
-    for (const facade of [
-      'system-record-v1.ts',
-      'system-record-objects-v1.ts',
-      'system-record-inventory-v1.ts',
-    ]) {
+  it('keeps the supported object and inventory facades explicit', () => {
+    for (const facade of ['system-record-objects-v1.ts', 'system-record-inventory-v1.ts']) {
       expect(source(facade)).not.toMatch(/export\s+\*/u);
     }
   });
 
+  it('composes the package facade only from the approved public subfacades', () => {
+    const facade = source('system-record-v1.ts');
+    const wildcardSources = [...facade.matchAll(/export\s+\*\s+from\s+'([^']+)'/gu)]
+      .map((match) => match[1]);
+    expect(wildcardSources).toEqual([
+      './system-record-limits-v1.js',
+      './system-record-objects-v1.js',
+      './agent-profile-projection-schema-v1.js',
+      './system-record-applied-state-v1.js',
+      './system-record-inventory-v1.js',
+      './system-record-wire-v1.js',
+    ]);
+    expect(facade).not.toMatch(/-internal\.js/u);
+  });
+
   it('keeps internal ownership units off compatibility facades', () => {
     for (const unit of [
-      'system-record-agent-profile-codecs-v1-internal.ts',
+      'system-record-agent-profile-primitives-v1-internal.ts',
+      'system-record-agent-profile-head-codec-v1-internal.ts',
+      'system-record-agent-profile-control-codecs-v1-internal.ts',
+      'system-record-agent-profile-evidence-codecs-v1-internal.ts',
+      'system-record-owned-subject-codecs-v1-internal.ts',
       'system-record-signatures-v1-internal.ts',
       'system-record-authority-summary-v1-internal.ts',
       'system-record-authority-v1-internal.ts',
@@ -45,9 +60,15 @@ describe('System Record V1 module ownership', () => {
   });
 
   it('keeps profile data codecs independent of signature verification', () => {
-    expect(source('system-record-agent-profile-codecs-v1-internal.ts')).not.toMatch(
-      /system-record-signatures-v1-internal/u,
-    );
+    for (const unit of [
+      'system-record-agent-profile-primitives-v1-internal.ts',
+      'system-record-agent-profile-head-codec-v1-internal.ts',
+      'system-record-agent-profile-control-codecs-v1-internal.ts',
+      'system-record-agent-profile-evidence-codecs-v1-internal.ts',
+      'system-record-owned-subject-codecs-v1-internal.ts',
+    ]) {
+      expect(source(unit)).not.toMatch(/system-record-signatures-v1-internal/u);
+    }
   });
 
   it('keeps wire and applied-state codecs off authority and closure implementations', () => {
